Convert Account model to using oneOf
Currently a bunch of this model's fields are specified with a list for
`type` field however according to the OAS spec thats invalid and we
need to be using one of the combo types to specify instead.

However the code generated when using oneOf is kinda crazy since it
generates these crazy union types to represent them even if said
target language supports union types or dynamic typing.

I'm going to convert these fields over to just string, as I believe
thats what is intended anyway and that the api will reject if not
strings, but leaving this commit here if we ever wanna just pull out
the conversion to oneOf fields real quick