Merge separated parts of Bank model to fix generated code
While the allOf approach these "*Resource" models have in the current
spec are technically correct by OAS standards, the code and docs that
get generated from it are not at all clear.

Not only do we generate useless models like "IdentifiedResource" that
never get used anywhere since the generators that support it just take
the fields and don't use a reference to an IdentifiedResource
instance (which is a good thing), certain generators just completely
ignore it and only generate the model with the non ref parts of the
model.

The kotlin generator for BankResource was just only generating the 2
basic fields and completely ignored the rest of the fields. Php one
was generating everything but the IdentifiedResource model was 100%
just dead code that we were generating everytime for no reason.